Returns ISO codes and names for all language or only those available in this package

See the relevant Wikipedia article for details on the language codes.

Usage

languages(available = TRUE)

Arguments

available

logical, whether to return only the languages supported in this package.

Value

A data frame with a row for each languages (only those supported if available is TRUE) and columns for the several ISO codes (639-2, 639-3, 639-1) and the name.

Details

Note that:

the ISO 639-1 code for mainland Chinese was changed to zh-cn.

A list of stop words in the variety of Chinese spoken in the island of Taiwan is accessible using the ISO 639-1 zh-tw or the name 'Chinese Taiwan'.

Ancient Greek has been assigned an artifact ISO 639-1 code (gr) because it had none. Its ISO 639-2 and 639-3 codes are both grc.

Examples

# Return all languages in the ISO 639-2/3 standard
languages()
#>    ISO639-2 ISO639-3 ISO639-1             name
#> 1       afr      afr       af        Afrikaans
#> 2       ara      ara       ar           Arabic
#> 3       hye      hye       hy         Armenian
#> 4       eus      eus       eu           Basque
#> 5       ben      ben       bn          Bengali
#> 6       bre      bre       br           Breton
#> 7       bul      bul       bg        Bulgarian
#> 8       cat      cat       ca          Catalan
#> 9       hrv      hrv       hr         Croatian
#> 10      ces      ces       cs            Czech
#> 11      dan      dan       da           Danish
#> 12      nld      nld       nl            Dutch
#> 13      eng      eng       en          English
#> 14      epo      epo       eo        Esperanto
#> 15      est      est       et         Estonian
#> 16      fin      fin       fi          Finnish
#> 17      fra      fra       fr           French
#> 18      glg      glg       gl         Galician
#> 19      deu      deu       de           German
#> 20      ell      ell       el    Greek, Modern
#> 21      guj      guj       gu         Gujarati
#> 22      hau      hau       ha            Hausa
#> 23      heb      heb       he           Hebrew
#> 24      hin      hin       hi            Hindi
#> 25      hun      hun       hu        Hungarian
#> 26      ind      ind       id       Indonesian
#> 27      gle      gle       ga            Irish
#> 28      ita      ita       it          Italian
#> 29      jpn      jpn       ja         Japanese
#> 30      kor      kor       ko           Korean
#> 31      kur      kur       ku          Kurdish
#> 32      lat      lat       la            Latin
#> 33      lit      lit       lt       Lithuanian
#> 34      lav      lav       lv          Latvian
#> 35      msa      msa       ms            Malay
#> 36      mar      mar       mr          Marathi
#> 37      nor      nor       no        Norwegian
#> 38      fas      fas       fa          Persian
#> 39      pol      pol       pl           Polish
#> 40      por      por       pt       Portuguese
#> 41      ron      ron       ro         Romanian
#> 42      rus      rus       ru          Russian
#> 43      slk      slk       sk           Slovak
#> 44      slv      slv       sl        Slovenian
#> 45      som      som       so           Somali
#> 46      sot      sot       st  Sotho, Southern
#> 47      spa      spa       es          Spanish
#> 48      swa      swa       sw          Swahili
#> 49      swe      swe       sv          Swedish
#> 50      tha      tha       th             Thai
#> 51      tgl      tgl       tl          Tagalog
#> 52      tur      tur       tr          Turkish
#> 53      ukr      ukr       uk        Ukrainian
#> 54      urd      urd       ur             Urdu
#> 55      vie      vie       vi       Vietnamese
#> 56      yor      yor       yo           Yoruba
#> 57      zul      zul       zu             Zulu
#> 58      grc      grc       gr   Greek, Ancient
#> 59      zho      zho    zh-cn Chinese Mandarin
#> 60       NA       NA    zh-tw   Chinese Taiwan
#> 61      aze      aze       az      Azerbaijani
#> 62      kaz      kaz       kk           Kazakh
#> 63      nep      nep       ne           Nepali
#> 64      tgk      tgk       tg            Tajik
#> 65     <NA>     <NA>     <NA>             <NA>