Collection of stopwords in multiple languages

This function returns stop words contained in the StopwordsISO repository.

Usage

stopwords(lang = "en")

Arguments

lang

Language for which to retrieve the stop word among those supported by StopwordISO. This parameters supports:

three-letter ISO 639-2/3 codes (e.g., 'eng');

two-letter ISO639-1 codes ('en');

names based ISO 639-2 codes ('English' or 'english') and their unambiguous substrings ('engl', 'engli', etc.).

Value

A character vector containing the stop words from the selected language as listed in the StopwordISO repository.

Examples

# They all return the correct list of stop words!

stopwords('German')
#>   [1] "a"               "ab"              "aber"            "ach"            
#>   [5] "acht"            "achte"           "achten"          "achter"         
#>   [9] "achtes"          "ag"              "alle"            "allein"         
#>  [13] "allem"           "allen"           "aller"           "allerdings"     
#>  [17] "alles"           "allgemeinen"     "als"             "also"           
#>  [21] "am"              "an"              "ander"           "andere"         
#>  [25] "anderem"         "anderen"         "anderer"         "anderes"        
#>  [29] "anderm"          "andern"          "anderr"          "anders"         
#>  [33] "au"              "auch"            "auf"             "aus"            
#>  [37] "ausser"          "außer"           "ausserdem"       "außerdem"       
#>  [41] "b"               "bald"            "bei"             "beide"          
#>  [45] "beiden"          "beim"            "beispiel"        "bekannt"        
#>  [49] "bereits"         "besonders"       "besser"          "besten"         
#>  [53] "bin"             "bis"             "bisher"          "bist"           
#>  [57] "c"               "d"               "d.h"             "da"             
#>  [61] "dabei"           "dadurch"         "dafür"           "dagegen"        
#>  [65] "daher"           "dahin"           "dahinter"        "damals"         
#>  [69] "damit"           "danach"          "daneben"         "dank"           
#>  [73] "dann"            "daran"           "darauf"          "daraus"         
#>  [77] "darf"            "darfst"          "darin"           "darüber"        
#>  [81] "darum"           "darunter"        "das"             "dasein"         
#>  [85] "daselbst"        "dass"            "daß"             "dasselbe"       
#>  [89] "davon"           "davor"           "dazu"            "dazwischen"     
#>  [93] "dein"            "deine"           "deinem"          "deinen"         
#>  [97] "deiner"          "deines"          "dem"             "dementsprechend"
#> [101] "demgegenüber"    "demgemäss"       "demgemäß"        "demselben"      
#> [105] "demzufolge"      "den"             "denen"           "denn"           
#> [109] "denselben"       "der"             "deren"           "derer"          
#> [113] "derjenige"       "derjenigen"      "dermassen"       "dermaßen"       
#> [117] "derselbe"        "derselben"       "des"             "deshalb"        
#> [121] "desselben"       "dessen"          "deswegen"        "dich"           
#> [125] "die"             "diejenige"       "diejenigen"      "dies"           
#> [129] "diese"           "dieselbe"        "dieselben"       "diesem"         
#> [133] "diesen"          "dieser"          "dieses"          "dir"            
#> [137] "doch"            "dort"            "drei"            "drin"           
#> [141] "dritte"          "dritten"         "dritter"         "drittes"        
#> [145] "du"              "durch"           "durchaus"        "dürfen"         
#> [149] "dürft"           "durfte"          "durften"         "e"              
#> [153] "eben"            "ebenso"          "ehrlich"         "ei"             
#> [157] "ei,"             "eigen"           "eigene"          "eigenen"        
#> [161] "eigener"         "eigenes"         "ein"             "einander"       
#> [165] "eine"            "einem"           "einen"           "einer"          
#> [169] "eines"           "einig"           "einige"          "einigem"        
#> [173] "einigen"         "einiger"         "einiges"         "einmal"         
#> [177] "eins"            "elf"             "en"              "ende"           
#> [181] "endlich"         "entweder"        "er"              "ernst"          
#> [185] "erst"            "erste"           "ersten"          "erster"         
#> [189] "erstes"          "es"              "etwa"            "etwas"          
#> [193] "euch"            "euer"            "eure"            "eurem"          
#> [197] "euren"           "eurer"           "eures"           "f"              
#> [201] "folgende"        "früher"          "fünf"            "fünfte"         
#> [205] "fünften"         "fünfter"         "fünftes"         "für"            
#> [209] "g"               "gab"             "ganz"            "ganze"          
#> [213] "ganzen"          "ganzer"          "ganzes"          "gar"            
#> [217] "gedurft"         "gegen"           "gegenüber"       "gehabt"         
#> [221] "gehen"           "geht"            "gekannt"         "gekonnt"        
#> [225] "gemacht"         "gemocht"         "gemusst"         "genug"          
#> [229] "gerade"          "gern"            "gesagt"          "geschweige"     
#> [233] "gewesen"         "gewollt"         "geworden"        "gibt"           
#> [237] "ging"            "gleich"          "gott"            "gross"          
#> [241] "groß"            "grosse"          "große"           "grossen"        
#> [245] "großen"          "grosser"         "großer"          "grosses"        
#> [249] "großes"          "gut"             "gute"            "guter"          
#> [253] "gutes"           "h"               "hab"             "habe"           
#> [257] "haben"           "habt"            "hast"            "hat"            
#> [261] "hatte"           "hätte"           "hatten"          "hätten"         
#> [265] "hattest"         "hattet"          "heisst"          "her"            
#> [269] "heute"           "hier"            "hin"             "hinter"         
#> [273] "hoch"            "i"               "ich"             "ihm"            
#> [277] "ihn"             "ihnen"           "ihr"             "ihre"           
#> [281] "ihrem"           "ihren"           "ihrer"           "ihres"          
#> [285] "im"              "immer"           "in"              "indem"          
#> [289] "infolgedessen"   "ins"             "irgend"          "ist"            
#> [293] "j"               "ja"              "jahr"            "jahre"          
#> [297] "jahren"          "je"              "jede"            "jedem"          
#> [301] "jeden"           "jeder"           "jedermann"       "jedermanns"     
#> [305] "jedes"           "jedoch"          "jemand"          "jemandem"       
#> [309] "jemanden"        "jene"            "jenem"           "jenen"          
#> [313] "jener"           "jenes"           "jetzt"           "k"              
#> [317] "kam"             "kann"            "kannst"          "kaum"           
#> [321] "kein"            "keine"           "keinem"          "keinen"         
#> [325] "keiner"          "keines"          "kleine"          "kleinen"        
#> [329] "kleiner"         "kleines"         "kommen"          "kommt"          
#> [333] "können"          "könnt"           "konnte"          "könnte"         
#> [337] "konnten"         "kurz"            "l"               "lang"           
#> [341] "lange"           "leicht"          "leide"           "lieber"         
#> [345] "los"             "m"               "machen"          "macht"          
#> [349] "machte"          "mag"             "magst"           "mahn"           
#> [353] "mal"             "man"             "manche"          "manchem"        
#> [357] "manchen"         "mancher"         "manches"         "mann"           
#> [361] "mehr"            "mein"            "meine"           "meinem"         
#> [365] "meinen"          "meiner"          "meines"          "mensch"         
#> [369] "menschen"        "mich"            "mir"             "mit"            
#> [373] "mittel"          "mochte"          "möchte"          "mochten"        
#> [377] "mögen"           "möglich"         "mögt"            "morgen"         
#> [381] "muss"            "muß"             "müssen"          "musst"          
#> [385] "mußt"            "müsst"           "müßt"            "musste"         
#> [389] "mussten"         "n"               "na"              "nach"           
#> [393] "nachdem"         "nahm"            "natürlich"       "neben"          
#> [397] "nein"            "neue"            "neuen"           "neun"           
#> [401] "neunte"          "neunten"         "neunter"         "neuntes"        
#> [405] "nicht"           "nichts"          "nie"             "niemand"        
#> [409] "niemandem"       "niemanden"       "noch"            "nun"            
#> [413] "nur"             "o"               "ob"              "oben"           
#> [417] "oder"            "offen"           "oft"             "ohne"           
#> [421] "ordnung"         "p"               "q"               "r"              
#> [425] "recht"           "rechte"          "rechten"         "rechter"        
#> [429] "rechtes"         "richtig"         "rund"            "s"              
#> [433] "sa"              "sache"           "sagt"            "sagte"          
#> [437] "sah"             "satt"            "schlecht"        "schluss"        
#> [441] "schon"           "sechs"           "sechste"         "sechsten"       
#> [445] "sechster"        "sechstes"        "sehr"            "sei"            
#> [449] "seid"            "seien"           "sein"            "seine"          
#> [453] "seinem"          "seinen"          "seiner"          "seines"         
#> [457] "seit"            "seitdem"         "selbst"          "sich"           
#> [461] "sie"             "sieben"          "siebente"        "siebenten"      
#> [465] "siebenter"       "siebentes"       "sind"            "so"             
#> [469] "solang"          "solche"          "solchem"         "solchen"        
#> [473] "solcher"         "solches"         "soll"            "sollen"         
#> [477] "sollst"          "sollt"           "sollte"          "sollten"        
#> [481] "sondern"         "sonst"           "soweit"          "sowie"          
#> [485] "später"          "startseite"      "statt"           "steht"          
#> [489] "suche"           "t"               "tag"             "tage"           
#> [493] "tagen"           "tat"             "teil"            "tel"            
#> [497] "tritt"           "trotzdem"        "tun"             "u"              
#> [501] "über"            "überhaupt"       "übrigens"        "uhr"            
#> [505] "um"              "und"             "uns"             "unse"           
#> [509] "unsem"           "unsen"           "unser"           "unsere"         
#> [513] "unserem"         "unseren"         "unserer"         "unseres"        
#> [517] "unses"           "unter"           "v"               "vergangenen"    
#> [521] "viel"            "viele"           "vielem"          "vielen"         
#> [525] "vielleicht"      "vier"            "vierte"          "vierten"        
#> [529] "vierter"         "viertes"         "vom"             "von"            
#> [533] "vor"             "w"               "wahr"            "während"        
#> [537] "währenddem"      "währenddessen"   "wann"            "war"            
#> [541] "wäre"            "waren"           "warst"           "wart"           
#> [545] "warum"           "was"             "weg"             "wegen"          
#> [549] "weil"            "weit"            "weiter"          "weitere"        
#> [553] "weiteren"        "weiteres"        "welche"          "welchem"        
#> [557] "welchen"         "welcher"         "welches"         "wem"            
#> [561] "wen"             "wenig"           "wenige"          "weniger"        
#> [565] "weniges"         "wenigstens"      "wenn"            "wer"            
#> [569] "werde"           "werden"          "werdet"          "weshalb"        
#> [573] "wessen"          "wie"             "wieder"          "wieso"          
#> [577] "will"            "willst"          "wir"             "wird"           
#> [581] "wirklich"        "wirst"           "wissen"          "wo"             
#> [585] "woher"           "wohin"           "wohl"            "wollen"         
#> [589] "wollt"           "wollte"          "wollten"         "worden"         
#> [593] "wurde"           "würde"           "wurden"          "würden"         
#> [597] "x"               "y"               "z"               "z.b"            
#> [601] "zehn"            "zehnte"          "zehnten"         "zehnter"        
#> [605] "zehntes"         "zeit"            "zu"              "zuerst"         
#> [609] "zugleich"        "zum"             "zunächst"        "zur"            
#> [613] "zurück"          "zusammen"        "zwanzig"         "zwar"           
#> [617] "zwei"            "zweite"          "zweiten"         "zweiter"        
#> [621] "zweites"         "zwischen"        "zwölf"          
stopwords('germ')
#>   [1] "a"               "ab"              "aber"            "ach"            
#>   [5] "acht"            "achte"           "achten"          "achter"         
#>   [9] "achtes"          "ag"              "alle"            "allein"         
#>  [13] "allem"           "allen"           "aller"           "allerdings"     
#>  [17] "alles"           "allgemeinen"     "als"             "also"           
#>  [21] "am"              "an"              "ander"           "andere"         
#>  [25] "anderem"         "anderen"         "anderer"         "anderes"        
#>  [29] "anderm"          "andern"          "anderr"          "anders"         
#>  [33] "au"              "auch"            "auf"             "aus"            
#>  [37] "ausser"          "außer"           "ausserdem"       "außerdem"       
#>  [41] "b"               "bald"            "bei"             "beide"          
#>  [45] "beiden"          "beim"            "beispiel"        "bekannt"        
#>  [49] "bereits"         "besonders"       "besser"          "besten"         
#>  [53] "bin"             "bis"             "bisher"          "bist"           
#>  [57] "c"               "d"               "d.h"             "da"             
#>  [61] "dabei"           "dadurch"         "dafür"           "dagegen"        
#>  [65] "daher"           "dahin"           "dahinter"        "damals"         
#>  [69] "damit"           "danach"          "daneben"         "dank"           
#>  [73] "dann"            "daran"           "darauf"          "daraus"         
#>  [77] "darf"            "darfst"          "darin"           "darüber"        
#>  [81] "darum"           "darunter"        "das"             "dasein"         
#>  [85] "daselbst"        "dass"            "daß"             "dasselbe"       
#>  [89] "davon"           "davor"           "dazu"            "dazwischen"     
#>  [93] "dein"            "deine"           "deinem"          "deinen"         
#>  [97] "deiner"          "deines"          "dem"             "dementsprechend"
#> [101] "demgegenüber"    "demgemäss"       "demgemäß"        "demselben"      
#> [105] "demzufolge"      "den"             "denen"           "denn"           
#> [109] "denselben"       "der"             "deren"           "derer"          
#> [113] "derjenige"       "derjenigen"      "dermassen"       "dermaßen"       
#> [117] "derselbe"        "derselben"       "des"             "deshalb"        
#> [121] "desselben"       "dessen"          "deswegen"        "dich"           
#> [125] "die"             "diejenige"       "diejenigen"      "dies"           
#> [129] "diese"           "dieselbe"        "dieselben"       "diesem"         
#> [133] "diesen"          "dieser"          "dieses"          "dir"            
#> [137] "doch"            "dort"            "drei"            "drin"           
#> [141] "dritte"          "dritten"         "dritter"         "drittes"        
#> [145] "du"              "durch"           "durchaus"        "dürfen"         
#> [149] "dürft"           "durfte"          "durften"         "e"              
#> [153] "eben"            "ebenso"          "ehrlich"         "ei"             
#> [157] "ei,"             "eigen"           "eigene"          "eigenen"        
#> [161] "eigener"         "eigenes"         "ein"             "einander"       
#> [165] "eine"            "einem"           "einen"           "einer"          
#> [169] "eines"           "einig"           "einige"          "einigem"        
#> [173] "einigen"         "einiger"         "einiges"         "einmal"         
#> [177] "eins"            "elf"             "en"              "ende"           
#> [181] "endlich"         "entweder"        "er"              "ernst"          
#> [185] "erst"            "erste"           "ersten"          "erster"         
#> [189] "erstes"          "es"              "etwa"            "etwas"          
#> [193] "euch"            "euer"            "eure"            "eurem"          
#> [197] "euren"           "eurer"           "eures"           "f"              
#> [201] "folgende"        "früher"          "fünf"            "fünfte"         
#> [205] "fünften"         "fünfter"         "fünftes"         "für"            
#> [209] "g"               "gab"             "ganz"            "ganze"          
#> [213] "ganzen"          "ganzer"          "ganzes"          "gar"            
#> [217] "gedurft"         "gegen"           "gegenüber"       "gehabt"         
#> [221] "gehen"           "geht"            "gekannt"         "gekonnt"        
#> [225] "gemacht"         "gemocht"         "gemusst"         "genug"          
#> [229] "gerade"          "gern"            "gesagt"          "geschweige"     
#> [233] "gewesen"         "gewollt"         "geworden"        "gibt"           
#> [237] "ging"            "gleich"          "gott"            "gross"          
#> [241] "groß"            "grosse"          "große"           "grossen"        
#> [245] "großen"          "grosser"         "großer"          "grosses"        
#> [249] "großes"          "gut"             "gute"            "guter"          
#> [253] "gutes"           "h"               "hab"             "habe"           
#> [257] "haben"           "habt"            "hast"            "hat"            
#> [261] "hatte"           "hätte"           "hatten"          "hätten"         
#> [265] "hattest"         "hattet"          "heisst"          "her"            
#> [269] "heute"           "hier"            "hin"             "hinter"         
#> [273] "hoch"            "i"               "ich"             "ihm"            
#> [277] "ihn"             "ihnen"           "ihr"             "ihre"           
#> [281] "ihrem"           "ihren"           "ihrer"           "ihres"          
#> [285] "im"              "immer"           "in"              "indem"          
#> [289] "infolgedessen"   "ins"             "irgend"          "ist"            
#> [293] "j"               "ja"              "jahr"            "jahre"          
#> [297] "jahren"          "je"              "jede"            "jedem"          
#> [301] "jeden"           "jeder"           "jedermann"       "jedermanns"     
#> [305] "jedes"           "jedoch"          "jemand"          "jemandem"       
#> [309] "jemanden"        "jene"            "jenem"           "jenen"          
#> [313] "jener"           "jenes"           "jetzt"           "k"              
#> [317] "kam"             "kann"            "kannst"          "kaum"           
#> [321] "kein"            "keine"           "keinem"          "keinen"         
#> [325] "keiner"          "keines"          "kleine"          "kleinen"        
#> [329] "kleiner"         "kleines"         "kommen"          "kommt"          
#> [333] "können"          "könnt"           "konnte"          "könnte"         
#> [337] "konnten"         "kurz"            "l"               "lang"           
#> [341] "lange"           "leicht"          "leide"           "lieber"         
#> [345] "los"             "m"               "machen"          "macht"          
#> [349] "machte"          "mag"             "magst"           "mahn"           
#> [353] "mal"             "man"             "manche"          "manchem"        
#> [357] "manchen"         "mancher"         "manches"         "mann"           
#> [361] "mehr"            "mein"            "meine"           "meinem"         
#> [365] "meinen"          "meiner"          "meines"          "mensch"         
#> [369] "menschen"        "mich"            "mir"             "mit"            
#> [373] "mittel"          "mochte"          "möchte"          "mochten"        
#> [377] "mögen"           "möglich"         "mögt"            "morgen"         
#> [381] "muss"            "muß"             "müssen"          "musst"          
#> [385] "mußt"            "müsst"           "müßt"            "musste"         
#> [389] "mussten"         "n"               "na"              "nach"           
#> [393] "nachdem"         "nahm"            "natürlich"       "neben"          
#> [397] "nein"            "neue"            "neuen"           "neun"           
#> [401] "neunte"          "neunten"         "neunter"         "neuntes"        
#> [405] "nicht"           "nichts"          "nie"             "niemand"        
#> [409] "niemandem"       "niemanden"       "noch"            "nun"            
#> [413] "nur"             "o"               "ob"              "oben"           
#> [417] "oder"            "offen"           "oft"             "ohne"           
#> [421] "ordnung"         "p"               "q"               "r"              
#> [425] "recht"           "rechte"          "rechten"         "rechter"        
#> [429] "rechtes"         "richtig"         "rund"            "s"              
#> [433] "sa"              "sache"           "sagt"            "sagte"          
#> [437] "sah"             "satt"            "schlecht"        "schluss"        
#> [441] "schon"           "sechs"           "sechste"         "sechsten"       
#> [445] "sechster"        "sechstes"        "sehr"            "sei"            
#> [449] "seid"            "seien"           "sein"            "seine"          
#> [453] "seinem"          "seinen"          "seiner"          "seines"         
#> [457] "seit"            "seitdem"         "selbst"          "sich"           
#> [461] "sie"             "sieben"          "siebente"        "siebenten"      
#> [465] "siebenter"       "siebentes"       "sind"            "so"             
#> [469] "solang"          "solche"          "solchem"         "solchen"        
#> [473] "solcher"         "solches"         "soll"            "sollen"         
#> [477] "sollst"          "sollt"           "sollte"          "sollten"        
#> [481] "sondern"         "sonst"           "soweit"          "sowie"          
#> [485] "später"          "startseite"      "statt"           "steht"          
#> [489] "suche"           "t"               "tag"             "tage"           
#> [493] "tagen"           "tat"             "teil"            "tel"            
#> [497] "tritt"           "trotzdem"        "tun"             "u"              
#> [501] "über"            "überhaupt"       "übrigens"        "uhr"            
#> [505] "um"              "und"             "uns"             "unse"           
#> [509] "unsem"           "unsen"           "unser"           "unsere"         
#> [513] "unserem"         "unseren"         "unserer"         "unseres"        
#> [517] "unses"           "unter"           "v"               "vergangenen"    
#> [521] "viel"            "viele"           "vielem"          "vielen"         
#> [525] "vielleicht"      "vier"            "vierte"          "vierten"        
#> [529] "vierter"         "viertes"         "vom"             "von"            
#> [533] "vor"             "w"               "wahr"            "während"        
#> [537] "währenddem"      "währenddessen"   "wann"            "war"            
#> [541] "wäre"            "waren"           "warst"           "wart"           
#> [545] "warum"           "was"             "weg"             "wegen"          
#> [549] "weil"            "weit"            "weiter"          "weitere"        
#> [553] "weiteren"        "weiteres"        "welche"          "welchem"        
#> [557] "welchen"         "welcher"         "welches"         "wem"            
#> [561] "wen"             "wenig"           "wenige"          "weniger"        
#> [565] "weniges"         "wenigstens"      "wenn"            "wer"            
#> [569] "werde"           "werden"          "werdet"          "weshalb"        
#> [573] "wessen"          "wie"             "wieder"          "wieso"          
#> [577] "will"            "willst"          "wir"             "wird"           
#> [581] "wirklich"        "wirst"           "wissen"          "wo"             
#> [585] "woher"           "wohin"           "wohl"            "wollen"         
#> [589] "wollt"           "wollte"          "wollten"         "worden"         
#> [593] "wurde"           "würde"           "wurden"          "würden"         
#> [597] "x"               "y"               "z"               "z.b"            
#> [601] "zehn"            "zehnte"          "zehnten"         "zehnter"        
#> [605] "zehntes"         "zeit"            "zu"              "zuerst"         
#> [609] "zugleich"        "zum"             "zunächst"        "zur"            
#> [613] "zurück"          "zusammen"        "zwanzig"         "zwar"           
#> [617] "zwei"            "zweite"          "zweiten"         "zweiter"        
#> [621] "zweites"         "zwischen"        "zwölf"          
stopwords('de')
#>   [1] "a"               "ab"              "aber"            "ach"            
#>   [5] "acht"            "achte"           "achten"          "achter"         
#>   [9] "achtes"          "ag"              "alle"            "allein"         
#>  [13] "allem"           "allen"           "aller"           "allerdings"     
#>  [17] "alles"           "allgemeinen"     "als"             "also"           
#>  [21] "am"              "an"              "ander"           "andere"         
#>  [25] "anderem"         "anderen"         "anderer"         "anderes"        
#>  [29] "anderm"          "andern"          "anderr"          "anders"         
#>  [33] "au"              "auch"            "auf"             "aus"            
#>  [37] "ausser"          "außer"           "ausserdem"       "außerdem"       
#>  [41] "b"               "bald"            "bei"             "beide"          
#>  [45] "beiden"          "beim"            "beispiel"        "bekannt"        
#>  [49] "bereits"         "besonders"       "besser"          "besten"         
#>  [53] "bin"             "bis"             "bisher"          "bist"           
#>  [57] "c"               "d"               "d.h"             "da"             
#>  [61] "dabei"           "dadurch"         "dafür"           "dagegen"        
#>  [65] "daher"           "dahin"           "dahinter"        "damals"         
#>  [69] "damit"           "danach"          "daneben"         "dank"           
#>  [73] "dann"            "daran"           "darauf"          "daraus"         
#>  [77] "darf"            "darfst"          "darin"           "darüber"        
#>  [81] "darum"           "darunter"        "das"             "dasein"         
#>  [85] "daselbst"        "dass"            "daß"             "dasselbe"       
#>  [89] "davon"           "davor"           "dazu"            "dazwischen"     
#>  [93] "dein"            "deine"           "deinem"          "deinen"         
#>  [97] "deiner"          "deines"          "dem"             "dementsprechend"
#> [101] "demgegenüber"    "demgemäss"       "demgemäß"        "demselben"      
#> [105] "demzufolge"      "den"             "denen"           "denn"           
#> [109] "denselben"       "der"             "deren"           "derer"          
#> [113] "derjenige"       "derjenigen"      "dermassen"       "dermaßen"       
#> [117] "derselbe"        "derselben"       "des"             "deshalb"        
#> [121] "desselben"       "dessen"          "deswegen"        "dich"           
#> [125] "die"             "diejenige"       "diejenigen"      "dies"           
#> [129] "diese"           "dieselbe"        "dieselben"       "diesem"         
#> [133] "diesen"          "dieser"          "dieses"          "dir"            
#> [137] "doch"            "dort"            "drei"            "drin"           
#> [141] "dritte"          "dritten"         "dritter"         "drittes"        
#> [145] "du"              "durch"           "durchaus"        "dürfen"         
#> [149] "dürft"           "durfte"          "durften"         "e"              
#> [153] "eben"            "ebenso"          "ehrlich"         "ei"             
#> [157] "ei,"             "eigen"           "eigene"          "eigenen"        
#> [161] "eigener"         "eigenes"         "ein"             "einander"       
#> [165] "eine"            "einem"           "einen"           "einer"          
#> [169] "eines"           "einig"           "einige"          "einigem"        
#> [173] "einigen"         "einiger"         "einiges"         "einmal"         
#> [177] "eins"            "elf"             "en"              "ende"           
#> [181] "endlich"         "entweder"        "er"              "ernst"          
#> [185] "erst"            "erste"           "ersten"          "erster"         
#> [189] "erstes"          "es"              "etwa"            "etwas"          
#> [193] "euch"            "euer"            "eure"            "eurem"          
#> [197] "euren"           "eurer"           "eures"           "f"              
#> [201] "folgende"        "früher"          "fünf"            "fünfte"         
#> [205] "fünften"         "fünfter"         "fünftes"         "für"            
#> [209] "g"               "gab"             "ganz"            "ganze"          
#> [213] "ganzen"          "ganzer"          "ganzes"          "gar"            
#> [217] "gedurft"         "gegen"           "gegenüber"       "gehabt"         
#> [221] "gehen"           "geht"            "gekannt"         "gekonnt"        
#> [225] "gemacht"         "gemocht"         "gemusst"         "genug"          
#> [229] "gerade"          "gern"            "gesagt"          "geschweige"     
#> [233] "gewesen"         "gewollt"         "geworden"        "gibt"           
#> [237] "ging"            "gleich"          "gott"            "gross"          
#> [241] "groß"            "grosse"          "große"           "grossen"        
#> [245] "großen"          "grosser"         "großer"          "grosses"        
#> [249] "großes"          "gut"             "gute"            "guter"          
#> [253] "gutes"           "h"               "hab"             "habe"           
#> [257] "haben"           "habt"            "hast"            "hat"            
#> [261] "hatte"           "hätte"           "hatten"          "hätten"         
#> [265] "hattest"         "hattet"          "heisst"          "her"            
#> [269] "heute"           "hier"            "hin"             "hinter"         
#> [273] "hoch"            "i"               "ich"             "ihm"            
#> [277] "ihn"             "ihnen"           "ihr"             "ihre"           
#> [281] "ihrem"           "ihren"           "ihrer"           "ihres"          
#> [285] "im"              "immer"           "in"              "indem"          
#> [289] "infolgedessen"   "ins"             "irgend"          "ist"            
#> [293] "j"               "ja"              "jahr"            "jahre"          
#> [297] "jahren"          "je"              "jede"            "jedem"          
#> [301] "jeden"           "jeder"           "jedermann"       "jedermanns"     
#> [305] "jedes"           "jedoch"          "jemand"          "jemandem"       
#> [309] "jemanden"        "jene"            "jenem"           "jenen"          
#> [313] "jener"           "jenes"           "jetzt"           "k"              
#> [317] "kam"             "kann"            "kannst"          "kaum"           
#> [321] "kein"            "keine"           "keinem"          "keinen"         
#> [325] "keiner"          "keines"          "kleine"          "kleinen"        
#> [329] "kleiner"         "kleines"         "kommen"          "kommt"          
#> [333] "können"          "könnt"           "konnte"          "könnte"         
#> [337] "konnten"         "kurz"            "l"               "lang"           
#> [341] "lange"           "leicht"          "leide"           "lieber"         
#> [345] "los"             "m"               "machen"          "macht"          
#> [349] "machte"          "mag"             "magst"           "mahn"           
#> [353] "mal"             "man"             "manche"          "manchem"        
#> [357] "manchen"         "mancher"         "manches"         "mann"           
#> [361] "mehr"            "mein"            "meine"           "meinem"         
#> [365] "meinen"          "meiner"          "meines"          "mensch"         
#> [369] "menschen"        "mich"            "mir"             "mit"            
#> [373] "mittel"          "mochte"          "möchte"          "mochten"        
#> [377] "mögen"           "möglich"         "mögt"            "morgen"         
#> [381] "muss"            "muß"             "müssen"          "musst"          
#> [385] "mußt"            "müsst"           "müßt"            "musste"         
#> [389] "mussten"         "n"               "na"              "nach"           
#> [393] "nachdem"         "nahm"            "natürlich"       "neben"          
#> [397] "nein"            "neue"            "neuen"           "neun"           
#> [401] "neunte"          "neunten"         "neunter"         "neuntes"        
#> [405] "nicht"           "nichts"          "nie"             "niemand"        
#> [409] "niemandem"       "niemanden"       "noch"            "nun"            
#> [413] "nur"             "o"               "ob"              "oben"           
#> [417] "oder"            "offen"           "oft"             "ohne"           
#> [421] "ordnung"         "p"               "q"               "r"              
#> [425] "recht"           "rechte"          "rechten"         "rechter"        
#> [429] "rechtes"         "richtig"         "rund"            "s"              
#> [433] "sa"              "sache"           "sagt"            "sagte"          
#> [437] "sah"             "satt"            "schlecht"        "schluss"        
#> [441] "schon"           "sechs"           "sechste"         "sechsten"       
#> [445] "sechster"        "sechstes"        "sehr"            "sei"            
#> [449] "seid"            "seien"           "sein"            "seine"          
#> [453] "seinem"          "seinen"          "seiner"          "seines"         
#> [457] "seit"            "seitdem"         "selbst"          "sich"           
#> [461] "sie"             "sieben"          "siebente"        "siebenten"      
#> [465] "siebenter"       "siebentes"       "sind"            "so"             
#> [469] "solang"          "solche"          "solchem"         "solchen"        
#> [473] "solcher"         "solches"         "soll"            "sollen"         
#> [477] "sollst"          "sollt"           "sollte"          "sollten"        
#> [481] "sondern"         "sonst"           "soweit"          "sowie"          
#> [485] "später"          "startseite"      "statt"           "steht"          
#> [489] "suche"           "t"               "tag"             "tage"           
#> [493] "tagen"           "tat"             "teil"            "tel"            
#> [497] "tritt"           "trotzdem"        "tun"             "u"              
#> [501] "über"            "überhaupt"       "übrigens"        "uhr"            
#> [505] "um"              "und"             "uns"             "unse"           
#> [509] "unsem"           "unsen"           "unser"           "unsere"         
#> [513] "unserem"         "unseren"         "unserer"         "unseres"        
#> [517] "unses"           "unter"           "v"               "vergangenen"    
#> [521] "viel"            "viele"           "vielem"          "vielen"         
#> [525] "vielleicht"      "vier"            "vierte"          "vierten"        
#> [529] "vierter"         "viertes"         "vom"             "von"            
#> [533] "vor"             "w"               "wahr"            "während"        
#> [537] "währenddem"      "währenddessen"   "wann"            "war"            
#> [541] "wäre"            "waren"           "warst"           "wart"           
#> [545] "warum"           "was"             "weg"             "wegen"          
#> [549] "weil"            "weit"            "weiter"          "weitere"        
#> [553] "weiteren"        "weiteres"        "welche"          "welchem"        
#> [557] "welchen"         "welcher"         "welches"         "wem"            
#> [561] "wen"             "wenig"           "wenige"          "weniger"        
#> [565] "weniges"         "wenigstens"      "wenn"            "wer"            
#> [569] "werde"           "werden"          "werdet"          "weshalb"        
#> [573] "wessen"          "wie"             "wieder"          "wieso"          
#> [577] "will"            "willst"          "wir"             "wird"           
#> [581] "wirklich"        "wirst"           "wissen"          "wo"             
#> [585] "woher"           "wohin"           "wohl"            "wollen"         
#> [589] "wollt"           "wollte"          "wollten"         "worden"         
#> [593] "wurde"           "würde"           "wurden"          "würden"         
#> [597] "x"               "y"               "z"               "z.b"            
#> [601] "zehn"            "zehnte"          "zehnten"         "zehnter"        
#> [605] "zehntes"         "zeit"            "zu"              "zuerst"         
#> [609] "zugleich"        "zum"             "zunächst"        "zur"            
#> [613] "zurück"          "zusammen"        "zwanzig"         "zwar"           
#> [617] "zwei"            "zweite"          "zweiten"         "zweiter"        
#> [621] "zweites"         "zwischen"        "zwölf"          
stopwords('deu')
#>   [1] "a"               "ab"              "aber"            "ach"            
#>   [5] "acht"            "achte"           "achten"          "achter"         
#>   [9] "achtes"          "ag"              "alle"            "allein"         
#>  [13] "allem"           "allen"           "aller"           "allerdings"     
#>  [17] "alles"           "allgemeinen"     "als"             "also"           
#>  [21] "am"              "an"              "ander"           "andere"         
#>  [25] "anderem"         "anderen"         "anderer"         "anderes"        
#>  [29] "anderm"          "andern"          "anderr"          "anders"         
#>  [33] "au"              "auch"            "auf"             "aus"            
#>  [37] "ausser"          "außer"           "ausserdem"       "außerdem"       
#>  [41] "b"               "bald"            "bei"             "beide"          
#>  [45] "beiden"          "beim"            "beispiel"        "bekannt"        
#>  [49] "bereits"         "besonders"       "besser"          "besten"         
#>  [53] "bin"             "bis"             "bisher"          "bist"           
#>  [57] "c"               "d"               "d.h"             "da"             
#>  [61] "dabei"           "dadurch"         "dafür"           "dagegen"        
#>  [65] "daher"           "dahin"           "dahinter"        "damals"         
#>  [69] "damit"           "danach"          "daneben"         "dank"           
#>  [73] "dann"            "daran"           "darauf"          "daraus"         
#>  [77] "darf"            "darfst"          "darin"           "darüber"        
#>  [81] "darum"           "darunter"        "das"             "dasein"         
#>  [85] "daselbst"        "dass"            "daß"             "dasselbe"       
#>  [89] "davon"           "davor"           "dazu"            "dazwischen"     
#>  [93] "dein"            "deine"           "deinem"          "deinen"         
#>  [97] "deiner"          "deines"          "dem"             "dementsprechend"
#> [101] "demgegenüber"    "demgemäss"       "demgemäß"        "demselben"      
#> [105] "demzufolge"      "den"             "denen"           "denn"           
#> [109] "denselben"       "der"             "deren"           "derer"          
#> [113] "derjenige"       "derjenigen"      "dermassen"       "dermaßen"       
#> [117] "derselbe"        "derselben"       "des"             "deshalb"        
#> [121] "desselben"       "dessen"          "deswegen"        "dich"           
#> [125] "die"             "diejenige"       "diejenigen"      "dies"           
#> [129] "diese"           "dieselbe"        "dieselben"       "diesem"         
#> [133] "diesen"          "dieser"          "dieses"          "dir"            
#> [137] "doch"            "dort"            "drei"            "drin"           
#> [141] "dritte"          "dritten"         "dritter"         "drittes"        
#> [145] "du"              "durch"           "durchaus"        "dürfen"         
#> [149] "dürft"           "durfte"          "durften"         "e"              
#> [153] "eben"            "ebenso"          "ehrlich"         "ei"             
#> [157] "ei,"             "eigen"           "eigene"          "eigenen"        
#> [161] "eigener"         "eigenes"         "ein"             "einander"       
#> [165] "eine"            "einem"           "einen"           "einer"          
#> [169] "eines"           "einig"           "einige"          "einigem"        
#> [173] "einigen"         "einiger"         "einiges"         "einmal"         
#> [177] "eins"            "elf"             "en"              "ende"           
#> [181] "endlich"         "entweder"        "er"              "ernst"          
#> [185] "erst"            "erste"           "ersten"          "erster"         
#> [189] "erstes"          "es"              "etwa"            "etwas"          
#> [193] "euch"            "euer"            "eure"            "eurem"          
#> [197] "euren"           "eurer"           "eures"           "f"              
#> [201] "folgende"        "früher"          "fünf"            "fünfte"         
#> [205] "fünften"         "fünfter"         "fünftes"         "für"            
#> [209] "g"               "gab"             "ganz"            "ganze"          
#> [213] "ganzen"          "ganzer"          "ganzes"          "gar"            
#> [217] "gedurft"         "gegen"           "gegenüber"       "gehabt"         
#> [221] "gehen"           "geht"            "gekannt"         "gekonnt"        
#> [225] "gemacht"         "gemocht"         "gemusst"         "genug"          
#> [229] "gerade"          "gern"            "gesagt"          "geschweige"     
#> [233] "gewesen"         "gewollt"         "geworden"        "gibt"           
#> [237] "ging"            "gleich"          "gott"            "gross"          
#> [241] "groß"            "grosse"          "große"           "grossen"        
#> [245] "großen"          "grosser"         "großer"          "grosses"        
#> [249] "großes"          "gut"             "gute"            "guter"          
#> [253] "gutes"           "h"               "hab"             "habe"           
#> [257] "haben"           "habt"            "hast"            "hat"            
#> [261] "hatte"           "hätte"           "hatten"          "hätten"         
#> [265] "hattest"         "hattet"          "heisst"          "her"            
#> [269] "heute"           "hier"            "hin"             "hinter"         
#> [273] "hoch"            "i"               "ich"             "ihm"            
#> [277] "ihn"             "ihnen"           "ihr"             "ihre"           
#> [281] "ihrem"           "ihren"           "ihrer"           "ihres"          
#> [285] "im"              "immer"           "in"              "indem"          
#> [289] "infolgedessen"   "ins"             "irgend"          "ist"            
#> [293] "j"               "ja"              "jahr"            "jahre"          
#> [297] "jahren"          "je"              "jede"            "jedem"          
#> [301] "jeden"           "jeder"           "jedermann"       "jedermanns"     
#> [305] "jedes"           "jedoch"          "jemand"          "jemandem"       
#> [309] "jemanden"        "jene"            "jenem"           "jenen"          
#> [313] "jener"           "jenes"           "jetzt"           "k"              
#> [317] "kam"             "kann"            "kannst"          "kaum"           
#> [321] "kein"            "keine"           "keinem"          "keinen"         
#> [325] "keiner"          "keines"          "kleine"          "kleinen"        
#> [329] "kleiner"         "kleines"         "kommen"          "kommt"          
#> [333] "können"          "könnt"           "konnte"          "könnte"         
#> [337] "konnten"         "kurz"            "l"               "lang"           
#> [341] "lange"           "leicht"          "leide"           "lieber"         
#> [345] "los"             "m"               "machen"          "macht"          
#> [349] "machte"          "mag"             "magst"           "mahn"           
#> [353] "mal"             "man"             "manche"          "manchem"        
#> [357] "manchen"         "mancher"         "manches"         "mann"           
#> [361] "mehr"            "mein"            "meine"           "meinem"         
#> [365] "meinen"          "meiner"          "meines"          "mensch"         
#> [369] "menschen"        "mich"            "mir"             "mit"            
#> [373] "mittel"          "mochte"          "möchte"          "mochten"        
#> [377] "mögen"           "möglich"         "mögt"            "morgen"         
#> [381] "muss"            "muß"             "müssen"          "musst"          
#> [385] "mußt"            "müsst"           "müßt"            "musste"         
#> [389] "mussten"         "n"               "na"              "nach"           
#> [393] "nachdem"         "nahm"            "natürlich"       "neben"          
#> [397] "nein"            "neue"            "neuen"           "neun"           
#> [401] "neunte"          "neunten"         "neunter"         "neuntes"        
#> [405] "nicht"           "nichts"          "nie"             "niemand"        
#> [409] "niemandem"       "niemanden"       "noch"            "nun"            
#> [413] "nur"             "o"               "ob"              "oben"           
#> [417] "oder"            "offen"           "oft"             "ohne"           
#> [421] "ordnung"         "p"               "q"               "r"              
#> [425] "recht"           "rechte"          "rechten"         "rechter"        
#> [429] "rechtes"         "richtig"         "rund"            "s"              
#> [433] "sa"              "sache"           "sagt"            "sagte"          
#> [437] "sah"             "satt"            "schlecht"        "schluss"        
#> [441] "schon"           "sechs"           "sechste"         "sechsten"       
#> [445] "sechster"        "sechstes"        "sehr"            "sei"            
#> [449] "seid"            "seien"           "sein"            "seine"          
#> [453] "seinem"          "seinen"          "seiner"          "seines"         
#> [457] "seit"            "seitdem"         "selbst"          "sich"           
#> [461] "sie"             "sieben"          "siebente"        "siebenten"      
#> [465] "siebenter"       "siebentes"       "sind"            "so"             
#> [469] "solang"          "solche"          "solchem"         "solchen"        
#> [473] "solcher"         "solches"         "soll"            "sollen"         
#> [477] "sollst"          "sollt"           "sollte"          "sollten"        
#> [481] "sondern"         "sonst"           "soweit"          "sowie"          
#> [485] "später"          "startseite"      "statt"           "steht"          
#> [489] "suche"           "t"               "tag"             "tage"           
#> [493] "tagen"           "tat"             "teil"            "tel"            
#> [497] "tritt"           "trotzdem"        "tun"             "u"              
#> [501] "über"            "überhaupt"       "übrigens"        "uhr"            
#> [505] "um"              "und"             "uns"             "unse"           
#> [509] "unsem"           "unsen"           "unser"           "unsere"         
#> [513] "unserem"         "unseren"         "unserer"         "unseres"        
#> [517] "unses"           "unter"           "v"               "vergangenen"    
#> [521] "viel"            "viele"           "vielem"          "vielen"         
#> [525] "vielleicht"      "vier"            "vierte"          "vierten"        
#> [529] "vierter"         "viertes"         "vom"             "von"            
#> [533] "vor"             "w"               "wahr"            "während"        
#> [537] "währenddem"      "währenddessen"   "wann"            "war"            
#> [541] "wäre"            "waren"           "warst"           "wart"           
#> [545] "warum"           "was"             "weg"             "wegen"          
#> [549] "weil"            "weit"            "weiter"          "weitere"        
#> [553] "weiteren"        "weiteres"        "welche"          "welchem"        
#> [557] "welchen"         "welcher"         "welches"         "wem"            
#> [561] "wen"             "wenig"           "wenige"          "weniger"        
#> [565] "weniges"         "wenigstens"      "wenn"            "wer"            
#> [569] "werde"           "werden"          "werdet"          "weshalb"        
#> [573] "wessen"          "wie"             "wieder"          "wieso"          
#> [577] "will"            "willst"          "wir"             "wird"           
#> [581] "wirklich"        "wirst"           "wissen"          "wo"             
#> [585] "woher"           "wohin"           "wohl"            "wollen"         
#> [589] "wollt"           "wollte"          "wollten"         "worden"         
#> [593] "wurde"           "würde"           "wurden"          "würden"         
#> [597] "x"               "y"               "z"               "z.b"            
#> [601] "zehn"            "zehnte"          "zehnten"         "zehnter"        
#> [605] "zehntes"         "zeit"            "zu"              "zuerst"         
#> [609] "zugleich"        "zum"             "zunächst"        "zur"            
#> [613] "zurück"          "zusammen"        "zwanzig"         "zwar"           
#> [617] "zwei"            "zweite"          "zweiten"         "zweiter"        
#> [621] "zweites"         "zwischen"        "zwölf"